requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Script==2.0.6
Flask-Testing==0.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Markdown==2.6.11
markdown2==2.3.5
MarkupSafe==1.0
python-dotenv==0.7.1
requests==2.18.4
urllib3==1.22
Werkzeug==0.14.1
